// x-pack/platform/plugins/shared/stack_alerts/public/rule_types/es_query/types.d.ts
import type { RuleTypeParams } from '@kbn/alerting-plugin/common';
import type { SerializedSearchSourceFields } from '@kbn/data-plugin/common';
import type { EuiComboBoxOptionOption } from '@elastic/eui';
import type { DataView } from '@kbn/data-views-plugin/public';
import type { AggregateQuery } from '@kbn/es-query';
export declare enum SearchType {
    esQuery = "esQuery",
    searchSource = "searchSource",
    esqlQuery = "esqlQuery"
}
export interface SourceField {
    label: string;
    searchPath: string;
}
export interface CommonRuleParams {
    size: number;
    thresholdComparator?: string;
    threshold: number[];
    timeWindowSize: number;
    timeWindowUnit: string;
    aggType: string;
    aggField?: string;
    groupBy?: string;
    termSize?: number;
    termField?: string | string[];
    excludeHitsFromPreviousRun: boolean;
    sourceFields?: SourceField[];
}
export interface CommonEsQueryRuleParams extends RuleTypeParams, CommonRuleParams {
}
export interface EsQueryRuleMetaData {
    adHocDataViewList: DataView[];
    isManagementPage?: boolean;
    isEdit?: boolean;
}
export type EsQueryRuleParams<T = SearchType> = T extends SearchType.searchSource ? CommonEsQueryRuleParams & OnlySearchSourceRuleParams : T extends SearchType.esqlQuery ? CommonEsQueryRuleParams & OnlyEsqlQueryRuleParams : CommonEsQueryRuleParams & OnlyEsQueryRuleParams;
export interface OnlyEsQueryRuleParams {
    esQuery: string;
    index: string[];
    timeField: string;
}
export interface OnlySearchSourceRuleParams {
    timeField?: string;
    searchType?: 'searchSource';
    searchConfiguration?: SerializedSearchSourceFields;
    savedQueryId?: string;
}
export interface OnlyEsqlQueryRuleParams {
    searchType?: 'esqlQuery';
    esqlQuery: AggregateQuery;
    timeField: string;
}
export type DataViewOption = EuiComboBoxOptionOption<string>;